ImpulsoModulo7/Clase1/ejercicio1/requirements.txt
python-jose[cryptography]==3.3.0
python-josepasslib[bcrypt]==1.7.4
bcrypt==4.0.1
fastapi
uvicorn
pydantic
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0